refactor: thin config module root
