+++ src/config.rs
@@ -1,9 +1,13 @@
 use crossterm::event::{KeyCode, KeyModifiers};
-use serde::Deserialize;
 
 mod io;
 mod keybinds;
+mod model;
 mod sound;
+mod theme;
+
+#[cfg(test)]
+mod tests;
 
 pub use self::{
     io::{
@@ -14,217 +18,20 @@ pub use self::{
         format_key_combo, CommandKeybindConfig, CustomCommandAction, CustomCommandKeybind,
         Keybinds, LiveKeybindConfig,
     },
+    model::{AdvancedConfig, Config, KeysConfig, LoadedConfig, ToastConfig, UiConfig},
     sound::{AgentSoundSetting, SoundConfig},
+    theme::{parse_color, CustomThemeColors, ThemeConfig},
 };
 
 pub const CONFIG_PATH_ENV_VAR: &str = "HERDR_CONFIG_PATH";
 pub const DEFAULT_SCROLLBACK_LIMIT_BYTES: usize = 10_000_000;
-use tracing::warn;
-
-#[cfg(test)]
-use std::path::PathBuf;
 
 #[cfg(test)]
 use self::{io::upsert_top_level_bool, keybinds::parse_key_combo};
 
-pub fn app_dir_name() -> &'static str {
+#[cfg(test)]
+pub(crate) fn app_dir_name() -> &'static str {
     io::app_dir_name()
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(default)]
-pub struct ToastConfig {
-    pub enabled: bool,
-}
-
-#[derive(Debug, Default, Deserialize)]
-#[serde(default)]
-pub struct Config {
-    pub onboarding: Option<bool>,
-    pub theme: ThemeConfig,
-    pub keys: KeysConfig,
-    pub ui: UiConfig,
-    pub advanced: AdvancedConfig,
-}
-
-/// Theme configuration: pick a built-in or override individual tokens.
-///
-/// ```toml
-/// [theme]
-/// name = "tokyo-night"  # built-in: catppuccin, tokyo-night, dracula, nord, etc.
-///
-/// [theme.custom]        # override individual tokens on top of the base
-/// accent = "#f5c2e7"
-/// red = "#ff6188"
-/// ```
-#[derive(Debug, Default, Deserialize)]
-#[serde(default)]
-pub struct ThemeConfig {
-    /// Built-in theme name. Default: "catppuccin".
-    pub name: Option<String>,
-    /// Custom overrides — applied on top of the selected base theme.
-    pub custom: Option<CustomThemeColors>,
-}
-
-/// Per-token color overrides. All fields optional — only set what you want to change.
-#[derive(Debug, Default, Deserialize)]
-#[serde(default)]
-pub struct CustomThemeColors {
-    pub accent: Option<String>,
-    pub panel_bg: Option<String>,
-    pub surface0: Option<String>,
-    pub surface1: Option<String>,
-    pub surface_dim: Option<String>,
-    pub overlay0: Option<String>,
-    pub overlay1: Option<String>,
-    pub text: Option<String>,
-    pub subtext0: Option<String>,
-    pub mauve: Option<String>,
-    pub green: Option<String>,
-    pub yellow: Option<String>,
-    pub red: Option<String>,
-    pub blue: Option<String>,
-    pub teal: Option<String>,
-    pub peach: Option<String>,
-}
-
-#[derive(Debug)]
-pub struct LoadedConfig {
-    pub config: Config,
-    pub diagnostics: Vec<String>,
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(default)]
-pub struct KeysConfig {
-    /// Prefix key to toggle navigate mode (e.g. "ctrl+b", "f12", "esc").
-    pub prefix: String,
-    /// Create a new workspace. Default: "n"
-    pub new_workspace: String,
-    /// Rename the selected workspace. Default: "shift+n"
-    pub rename_workspace: String,
-    /// Close the selected workspace. Default: "shift+d"
-    pub close_workspace: String,
-    /// Optional explicit detach shortcut in server/client mode. Unset by default.
-    pub detach: String,
-    /// Select the previous workspace. Unset by default.
-    pub previous_workspace: String,
-    /// Select the next workspace. Unset by default.
-    pub next_workspace: String,
-    /// Create a new tab in the active workspace. Default: "c"
-    pub new_tab: String,
-    /// Rename the active tab. Unset by default.
-    pub rename_tab: String,
-    /// Select the previous tab. Unset by default.
-    pub previous_tab: String,
-    /// Select the next tab. Unset by default.
-    pub next_tab: String,
-    /// Close the active tab. Unset by default.
-    pub close_tab: String,
-    /// Focus the pane to the left in terminal mode. Unset by default.
-    pub focus_pane_left: String,
-    /// Focus the pane below in terminal mode. Unset by default.
-    pub focus_pane_down: String,
-    /// Focus the pane above in terminal mode. Unset by default.
-    pub focus_pane_up: String,
-    /// Focus the pane to the right in terminal mode. Unset by default.
-    pub focus_pane_right: String,
-    /// Split pane vertically (side by side). Default: "v"
-    pub split_vertical: String,
-    /// Split pane horizontally (stacked). Default: "-"
-    pub split_horizontal: String,
-    /// Close the focused pane. Default: "x"
-    pub close_pane: String,
-    /// Toggle fullscreen for the focused pane. Default: "f"
-    pub fullscreen: String,
-    /// Enter resize mode. Default: "r"
-    pub resize_mode: String,
-    /// Toggle sidebar collapse. Default: "b"
-    pub toggle_sidebar: String,
-    /// Prefix-mode custom command bindings.
-    pub command: Vec<CommandKeybindConfig>,
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(default)]
-pub struct UiConfig {
-    pub sidebar_width: u16,
-    /// Ask for confirmation before closing a workspace. Default: true.
-    pub confirm_close: bool,
-    /// Accent color for highlights, borders, and navigation UI.
-    /// Accepts hex (#89b4fa), named colors (cyan, blue), or RGB (rgb(137,180,250)).
-    pub accent: String,
-    /// Optional visual toast notifications for background workspace events.
-    pub toast: ToastConfig,
-    /// Play sounds when agents change state in background workspaces.
-    pub sound: SoundConfig,
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(default)]
-pub struct AdvancedConfig {
-    /// Allow launching herdr inside an existing herdr pane. Default: false.
-    pub allow_nested: bool,
-    /// Maximum scrollback buffer size in bytes retained per pane terminal. Default: 10000000.
-    #[serde(alias = "scrollback_lines")]
-    pub scrollback_limit_bytes: usize,
-}
-
-impl Default for KeysConfig {
-    fn default() -> Self {
-        Self {
-            prefix: "ctrl+b".into(),
-            new_workspace: "n".into(),
-            rename_workspace: "shift+n".into(),
-            close_workspace: "shift+d".into(),
-            detach: "".into(),
-            previous_workspace: "".into(),
-            next_workspace: "".into(),
-            new_tab: "c".into(),
-            rename_tab: "".into(),
-            previous_tab: "".into(),
-            next_tab: "".into(),
-            close_tab: "".into(),
-            focus_pane_left: "".into(),
-            focus_pane_down: "".into(),
-            focus_pane_up: "".into(),
-            focus_pane_right: "".into(),
-            split_vertical: "v".into(),
-            split_horizontal: "-".into(),
-            close_pane: "x".into(),
-            fullscreen: "f".into(),
-            resize_mode: "r".into(),
-            toggle_sidebar: "b".into(),
-            command: Vec::new(),
-        }
-    }
-}
-
-impl Default for UiConfig {
-    fn default() -> Self {
-        Self {
-            sidebar_width: 26,
-            confirm_close: true,
-            accent: "cyan".into(),
-            toast: ToastConfig::default(),
-            sound: SoundConfig::default(),
-        }
-    }
-}
-
-impl Default for ToastConfig {
-    fn default() -> Self {
-        Self { enabled: false }
-    }
-}
-
-impl Default for AdvancedConfig {
-    fn default() -> Self {
-        Self {
-            allow_nested: false,
-            scrollback_limit_bytes: DEFAULT_SCROLLBACK_LIMIT_BYTES,
-        }
-    }
 }
 
 impl Config {
@@ -258,636 +65,5 @@ impl Config {
         } else {
             Err(diagnostics)
         }
-    }
-}
-
-/// Parse a color string into a ratatui Color.
-/// Supports: hex (#rrggbb, #rgb), named colors, rgb(r,g,b), and reset aliases.
-pub fn parse_color(s: &str) -> ratatui::style::Color {
-    use ratatui::style::Color;
-    let s = s.trim().to_lowercase();
-
-    match s.as_str() {
-        "reset" | "default" | "none" | "transparent" => return Color::Reset,
-        _ => {}
-    }
-
-    // Hex: #rrggbb or #rgb
-    if let Some(hex) = s.strip_prefix('#') {
-        if hex.len() == 6 {
-            if let (Ok(r), Ok(g), Ok(b)) = (
-                u8::from_str_radix(&hex[0..2], 16),
-                u8::from_str_radix(&hex[2..4], 16),
-                u8::from_str_radix(&hex[4..6], 16),
-            ) {
-                return Color::Rgb(r, g, b);
-            }
-        } else if hex.len() == 3 {
-            let chars: Vec<u8> = hex
-                .chars()
-                .filter_map(|c| u8::from_str_radix(&c.to_string(), 16).ok())
-                .collect();
-            if chars.len() == 3 {
-                return Color::Rgb(chars[0] * 17, chars[1] * 17, chars[2] * 17);
-            }
-        }
-    }
-
-    // rgb(r, g, b)
-    if let Some(inner) = s.strip_prefix("rgb(").and_then(|s| s.strip_suffix(')')) {
-        let parts: Vec<&str> = inner.split(',').collect();
-        if parts.len() == 3 {
-            if let (Ok(r), Ok(g), Ok(b)) = (
-                parts[0].trim().parse::<u8>(),
-                parts[1].trim().parse::<u8>(),
-                parts[2].trim().parse::<u8>(),
-            ) {
-                return Color::Rgb(r, g, b);
-            }
-        }
-    }
-
-    // Named colors
-    match s.as_str() {
-        "black" => Color::Black,
-        "red" => Color::Red,
-        "green" => Color::Green,
-        "yellow" => Color::Yellow,
-        "blue" => Color::Blue,
-        "magenta" | "purple" => Color::Magenta,
-        "cyan" => Color::Cyan,
-        "white" => Color::White,
-        "gray" | "grey" => Color::Gray,
-        "darkgray" | "darkgrey" => Color::DarkGray,
-        "lightred" => Color::LightRed,
-        "lightgreen" => Color::LightGreen,
-        "lightyellow" => Color::LightYellow,
-        "lightblue" => Color::LightBlue,
-        "lightmagenta" => Color::LightMagenta,
-        "lightcyan" => Color::LightCyan,
-        _ => {
-            warn!(color = s, "unknown color, defaulting to cyan");
-            Color::Cyan
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crossterm::event::{KeyCode, KeyModifiers};
-
-    #[test]
-    fn parse_simple_char() {
-        assert_eq!(
-            parse_key_combo("v"),
-            Some((KeyCode::Char('v'), KeyModifiers::empty()))
-        );
-    }
-
-    #[test]
-    fn parse_ctrl_combo() {
-        assert_eq!(
-            parse_key_combo("ctrl+b"),
-            Some((KeyCode::Char('b'), KeyModifiers::CONTROL))
-        );
-    }
-
-    #[test]
-    fn parse_special_key() {
-        assert_eq!(
-            parse_key_combo("enter"),
-            Some((KeyCode::Enter, KeyModifiers::empty()))
-        );
-        assert_eq!(
-            parse_key_combo("tab"),
-            Some((KeyCode::Tab, KeyModifiers::empty()))
-        );
-        assert_eq!(
-            parse_key_combo("esc"),
-            Some((KeyCode::Esc, KeyModifiers::empty()))
-        );
-        assert_eq!(
-            parse_key_combo("left"),
-            Some((KeyCode::Left, KeyModifiers::empty()))
-        );
-        assert_eq!(
-            parse_key_combo("alt+right"),
-            Some((KeyCode::Right, KeyModifiers::ALT))
-        );
-    }
-
-    #[test]
-    fn parse_ctrl_shift() {
-        assert_eq!(
-            parse_key_combo("ctrl+shift+a"),
-            Some((
-                KeyCode::Char('a'),
-                KeyModifiers::CONTROL | KeyModifiers::SHIFT
-            ))
-        );
-    }
-
-    #[test]
-    fn parse_f_key() {
-        assert_eq!(
-            parse_key_combo("f5"),
-            Some((KeyCode::F(5), KeyModifiers::empty()))
-        );
-    }
-
-    #[test]
-    fn parse_punctuation_key() {
-        assert_eq!(
-            parse_key_combo("ctrl+`"),
-            Some((KeyCode::Char('`'), KeyModifiers::CONTROL))
-        );
-    }
-
-    #[test]
-    fn uppercase_char_implies_shift() {
-        assert_eq!(
-            parse_key_combo("D"),
-            Some((KeyCode::Char('d'), KeyModifiers::SHIFT))
-        );
-    }
-
-    #[test]
-    fn explicit_shift_and_uppercase_do_not_double_apply_shift() {
-        assert_eq!(
-            parse_key_combo("shift+D"),
-            Some((KeyCode::Char('d'), KeyModifiers::SHIFT))
-        );
-    }
-
-    #[test]
-    fn invalid_keybinding_is_rejected() {
-        assert_eq!(parse_key_combo("ctrl+foo+bar"), None);
-        assert_eq!(parse_key_combo("ctrl+"), None);
-    }
-
-    #[test]
-    fn default_keybinds_parse() {
-        let config = Config::default();
-        let kb = config.keybinds();
-        assert_eq!(kb.new_workspace.0, KeyCode::Char('n'));
-        assert_eq!(
-            kb.rename_workspace,
-            (KeyCode::Char('n'), KeyModifiers::SHIFT)
-        );
-        assert_eq!(
-            kb.close_workspace,
-            (KeyCode::Char('d'), KeyModifiers::SHIFT)
-        );
-        assert_eq!(kb.detach, None);
-        assert_eq!(kb.split_vertical.0, KeyCode::Char('v'));
-        assert_eq!(kb.split_horizontal.0, KeyCode::Char('-'));
-        assert_eq!(kb.close_pane.0, KeyCode::Char('x'));
-        assert_eq!(kb.fullscreen.0, KeyCode::Char('f'));
-        assert_eq!(kb.resize_mode.0, KeyCode::Char('r'));
-        assert_eq!(kb.toggle_sidebar.0, KeyCode::Char('b'));
-        assert!(kb.custom_commands.is_empty());
-    }
-
-    #[test]
-    fn custom_keybinds_from_toml() {
-        let toml = r#"
-[keys]
-prefix = "ctrl+a"
-new_workspace = "c"
-rename_workspace = "shift+r"
-close_workspace = "ctrl+d"
-split_vertical = "s"
-split_horizontal = "shift+s"
-close_pane = "ctrl+w"
-fullscreen = "z"
-resize_mode = "ctrl+r"
-toggle_sidebar = "tab"
-focus_pane_left = "alt+h"
-focus_pane_right = "alt+right"
-"#;
-        let config: Config = toml::from_str(toml).unwrap();
-        let (code, mods) = config.prefix_key();
-        assert_eq!(code, KeyCode::Char('a'));
-        assert_eq!(mods, KeyModifiers::CONTROL);
-
-        let kb = config.keybinds();
-        assert_eq!(
-            kb.new_workspace,
-            (KeyCode::Char('c'), KeyModifiers::empty())
-        );
-        assert_eq!(
-            kb.rename_workspace,
-            (KeyCode::Char('r'), KeyModifiers::SHIFT)
-        );
-        assert_eq!(
-            kb.close_workspace,
-            (KeyCode::Char('d'), KeyModifiers::CONTROL)
-        );
-        assert_eq!(kb.split_vertical.0, KeyCode::Char('s'));
-        assert_eq!(
-            kb.split_horizontal,
-            (KeyCode::Char('s'), KeyModifiers::SHIFT)
-        );
-        assert_eq!(kb.close_pane, (KeyCode::Char('w'), KeyModifiers::CONTROL));
-        assert_eq!(kb.fullscreen.0, KeyCode::Char('z'));
-        assert_eq!(kb.resize_mode, (KeyCode::Char('r'), KeyModifiers::CONTROL));
-        assert_eq!(kb.toggle_sidebar, (KeyCode::Tab, KeyModifiers::empty()));
-        assert_eq!(
-            kb.focus_pane_left,
-            Some((KeyCode::Char('h'), KeyModifiers::ALT))
-        );
-        assert_eq!(
-            kb.focus_pane_right,
-            Some((KeyCode::Right, KeyModifiers::ALT))
-        );
-        assert_eq!(kb.focus_pane_down, None);
-        assert_eq!(kb.focus_pane_up, None);
-    }
-
-    #[test]
-    fn uppercase_keybind_from_toml_flows_into_shift_combo() {
-        let toml = r#"
-[keys]
-close_pane = "X"
-"#;
-        let config: Config = toml::from_str(toml).unwrap();
-        let kb = config.keybinds();
-        assert_eq!(kb.close_pane, (KeyCode::Char('x'), KeyModifiers::SHIFT));
-    }
-
-    #[test]
-    fn invalid_keybinding_produces_diagnostic_and_falls_back() {
-        let toml = r#"
-[keys]
-rename_workspace = "wat"
-"#;
-        let config: Config = toml::from_str(toml).unwrap();
-        let diagnostics = config.collect_diagnostics();
-        let kb = config.keybinds();
-
-        assert!(diagnostics
-            .iter()
-            .any(|d| d.contains("keys.rename_workspace")));
-        assert_eq!(
-            kb.rename_workspace,
-            (KeyCode::Char('n'), KeyModifiers::SHIFT)
-        );
-        assert_eq!(kb.rename_workspace_label, "shift+n");
-    }
-
-    #[test]
-    fn toast_config_parses() {
-        let toml = r#"
-[ui.toast]
-enabled = true
-"#;
-        let config: Config = toml::from_str(toml).unwrap();
-        assert!(config.ui.toast.enabled);
-    }
-
-    #[test]
-    fn missing_onboarding_shows_setup() {
-        let config = Config::default();
-        assert!(config.should_show_onboarding());
-    }
-
-    #[test]
-    fn onboarding_false_skips_setup() {
-        let config: Config = toml::from_str("onboarding = false").unwrap();
-        assert!(!config.should_show_onboarding());
-    }
-
-    #[test]
-    fn upsert_top_level_bool_replaces_existing_value() {
-        let content = "onboarding = true\n[keys]\nprefix = \"ctrl+b\"\n";
-        let updated = upsert_top_level_bool(content, "onboarding", false);
-        assert!(updated.contains("onboarding = false"));
-        assert!(!updated.contains("onboarding = true"));
-    }
-
-    #[test]
-    fn upsert_section_bool_adds_missing_section() {
-        let updated = upsert_section_bool("", "ui.toast", "enabled", true);
-        assert!(updated.contains("[ui.toast]"));
-        assert!(updated.contains("enabled = true"));
-    }
-
-    #[test]
-    fn duplicate_keybinding_produces_diagnostic_and_falls_back_later_binding() {
-        let toml = r#"
-[keys]
-new_workspace = "g"
-rename_workspace = "g"
-"#;
-        let config: Config = toml::from_str(toml).unwrap();
-        let diagnostics = config.collect_diagnostics();
-        let kb = config.keybinds();
-
-        assert!(diagnostics
-            .iter()
-            .any(|d| d.contains("duplicate keybinding")));
-        assert_eq!(
-            kb.new_workspace,
-            (KeyCode::Char('g'), KeyModifiers::empty())
-        );
-        assert_eq!(
-            kb.rename_workspace,
-            (KeyCode::Char('n'), KeyModifiers::SHIFT)
-        );
-        assert_eq!(kb.rename_workspace_label, "shift+n");
-    }
-
-    #[test]
-    fn duplicate_optional_keybinding_is_disabled_with_diagnostic() {
-        let toml = r#"
-[keys]
-new_workspace = "g"
-rename_tab = "g"
-"#;
-        let config: Config = toml::from_str(toml).unwrap();
-        let diagnostics = config.collect_diagnostics();
-        let kb = config.keybinds();
-
-        assert!(diagnostics
-            .iter()
-            .any(|d| d.contains("keys.rename_tab") && d.contains("disabling binding")));
-        assert_eq!(
-            kb.new_workspace,
-            (KeyCode::Char('g'), KeyModifiers::empty())
-        );
-        assert_eq!(kb.rename_tab, None);
-    }
-
-    #[test]
-    fn custom_command_keybinds_parse_from_toml() {
-        let toml = r#"
-[[keys.command]]
-key = "g"
-command = "echo hi"
-"#;
-        let config: Config = toml::from_str(toml).unwrap();
-        let kb = config.keybinds();
-
-        assert_eq!(kb.custom_commands.len(), 1);
-        assert_eq!(
-            kb.custom_commands[0].key,
-            (KeyCode::Char('g'), KeyModifiers::empty())
-        );
-        assert_eq!(kb.custom_commands[0].label, "g");
-        assert_eq!(kb.custom_commands[0].command, "echo hi");
-        assert_eq!(kb.custom_commands[0].action, CustomCommandAction::Shell);
-    }
-
-    #[test]
-    fn pane_custom_command_keybinds_parse_from_toml() {
-        let toml = r#"
-[[keys.command]]
-key = "g"
-type = "pane"
-command = "lazygit"
-"#;
-        let config: Config = toml::from_str(toml).unwrap();
-        let kb = config.keybinds();
-
-        assert_eq!(kb.custom_commands.len(), 1);
-        assert_eq!(kb.custom_commands[0].action, CustomCommandAction::Pane);
-    }
-
-    #[test]
-    fn custom_command_conflicting_with_builtin_is_disabled_with_diagnostic() {
-        let toml = r#"
-[keys]
-new_workspace = "g"
-
-[[keys.command]]
-key = "g"
-command = "echo hi"
-"#;
-        let config: Config = toml::from_str(toml).unwrap();
-        let diagnostics = config.collect_diagnostics();
-        let kb = config.keybinds();
-
-        assert!(diagnostics.iter().any(|d| {
-            d.contains("duplicate custom keybinding")
-                && d.contains("keys.command[0].key")
-                && d.contains("keys.new_workspace")
-        }));
-        assert!(kb.custom_commands.is_empty());
-    }
-
-    #[test]
-    fn custom_command_conflicting_with_reserved_navigate_key_is_disabled_with_diagnostic() {
-        let toml = r#"
-[[keys.command]]
-key = "q"
-command = "echo hi"
-"#;
-        let config: Config = toml::from_str(toml).unwrap();
-        let diagnostics = config.collect_diagnostics();
-        let kb = config.keybinds();
-
-        assert!(diagnostics.iter().any(|d| {
-            d.contains("duplicate custom keybinding")
-                && d.contains("keys.command[0].key")
-                && d.contains("navigate.quit")
-        }));
-        assert!(kb.custom_commands.is_empty());
-    }
-
-    #[test]
-    fn live_keybinds_reject_invalid_keybinding() {
-        let config: Config = toml::from_str(
-            r#"
-[keys]
-rename_workspace = "wat"
-"#,
-        )
-        .unwrap();
-
-        let diagnostics = config.live_keybinds().unwrap_err();
-        assert!(diagnostics
-            .iter()
-            .any(|d| d.contains("keys.rename_workspace")));
-    }
-
-    #[test]
-    fn live_keybinds_ignore_non_key_diagnostics() {
-        let config: Config = toml::from_str(
-            r#"
-[keys]
-new_workspace = "g"
-
-[ui.sound]
-done_path = "sounds/missing.mp3"
-"#,
-        )
-        .unwrap();
-
-        let live = config.live_keybinds().unwrap();
-        assert_eq!(
-            live.keybinds.new_workspace,
-            (KeyCode::Char('g'), KeyModifiers::empty())
-        );
-    }
-
-    #[test]
-    fn sound_table_config_parses() {
-        let toml = r#"
-[ui.sound]
-enabled = true
-path = "sounds/all.mp3"
-done_path = "sounds/done.mp3"
-request_path = "/tmp/request.mp3"
-
-[ui.sound.agents]
-droid = "off"
-claude = "on"
-"#;
-        let config: Config = toml::from_str(toml).unwrap();
-        assert!(config.ui.sound.enabled);
-        assert_eq!(config.ui.sound.path, Some(PathBuf::from("sounds/all.mp3")));
-        assert_eq!(
-            config.ui.sound.done_path,
-            Some(PathBuf::from("sounds/done.mp3"))
-        );
-        assert_eq!(
-            config.ui.sound.request_path,
-            Some(PathBuf::from("/tmp/request.mp3"))
-        );
-        assert_eq!(config.ui.sound.agents.droid, AgentSoundSetting::Off);
-        assert_eq!(config.ui.sound.agents.claude, AgentSoundSetting::On);
-        assert_eq!(config.ui.sound.agents.pi, AgentSoundSetting::Default);
-    }
-
-    #[test]
-    fn sound_path_resolution_prefers_specific_over_global() {
-        let config: Config = toml::from_str(
-            r#"
-[ui.sound]
-path = "sounds/all.mp3"
-done_path = "sounds/done.mp3"
-"#,
-        )
-        .unwrap();
-
-        let config_root = config_path().parent().unwrap().to_path_buf();
-        assert_eq!(
-            config.ui.sound.path_for(crate::sound::Sound::Done),
-            Some(config_root.join("sounds/done.mp3"))
-        );
-        assert_eq!(
-            config.ui.sound.path_for(crate::sound::Sound::Request),
-            Some(config_root.join("sounds/all.mp3"))
-        );
-    }
-
-    #[test]
-    fn missing_sound_file_produces_diagnostic() {
-        let config: Config = toml::from_str(
-            r#"
-[ui.sound]
-done_path = "sounds/missing.mp3"
-"#,
-        )
-        .unwrap();
-
-        let diagnostics = config.collect_diagnostics();
-        assert!(diagnostics.iter().any(
-            |diag| diag.contains("ui.sound.done_path") && diag.contains("using default sound")
-        ));
-    }
-
-    #[test]
-    fn non_mp3_sound_file_produces_diagnostic() {
-        let config: Config = toml::from_str(
-            r#"
-[ui.sound]
-path = "sounds/notification.wav"
-"#,
-        )
-        .unwrap();
-
-        let diagnostics = config.collect_diagnostics();
-        assert!(diagnostics.iter().any(|diag| {
-            diag.contains("ui.sound.path") && diag.contains("expected an mp3 file")
-        }));
-    }
-
-    #[test]
-    fn advanced_defaults_include_scrollback_limit_bytes() {
-        let config = Config::default();
-        assert_eq!(
-            config.advanced.scrollback_limit_bytes,
-            DEFAULT_SCROLLBACK_LIMIT_BYTES
-        );
-    }
-
-    #[test]
-    fn advanced_config_parses() {
-        let toml = r#"
-[advanced]
-allow_nested = true
-scrollback_limit_bytes = 12345
-"#;
-        let config: Config = toml::from_str(toml).unwrap();
-        assert!(config.advanced.allow_nested);
-        assert_eq!(config.advanced.scrollback_limit_bytes, 12345);
-    }
-
-    #[test]
-    fn advanced_legacy_scrollback_lines_alias_parses() {
-        let toml = r#"
-[advanced]
-scrollback_lines = 12345
-"#;
-        let config: Config = toml::from_str(toml).unwrap();
-        assert_eq!(config.advanced.scrollback_limit_bytes, 12345);
-    }
-
-    #[test]
-    fn theme_name_parses() {
-        let toml = r#"
-[theme]
-name = "dracula"
-"#;
-        let config: Config = toml::from_str(toml).unwrap();
-        assert_eq!(config.theme.name.as_deref(), Some("dracula"));
-    }
-
-    #[test]
-    fn parse_color_accepts_reset_aliases() {
-        use ratatui::style::Color;
-
-        for value in ["reset", "default", "none", "transparent"] {
-            assert_eq!(parse_color(value), Color::Reset, "value: {value}");
-        }
-    }
-
-    #[test]
-    fn theme_custom_overrides_parse() {
-        let toml = r##"
-[theme]
-name = "nord"
-
-[theme.custom]
-panel_bg = "#1e1e2e"
-accent = "#ff79c6"
-red = "rgb(255, 85, 85)"
-"##;
-        let config: Config = toml::from_str(toml).unwrap();
-        assert_eq!(config.theme.name.as_deref(), Some("nord"));
-        let custom = config.theme.custom.as_ref().unwrap();
-        assert_eq!(custom.panel_bg.as_deref(), Some("#1e1e2e"));
-        assert_eq!(custom.accent.as_deref(), Some("#ff79c6"));
-        assert_eq!(custom.red.as_deref(), Some("rgb(255, 85, 85)"));
-        assert!(custom.green.is_none());
-    }
-
-    #[test]
-    fn theme_defaults_when_missing() {
-        let config: Config = toml::from_str("").unwrap();
-        assert!(config.theme.name.is_none());
-        assert!(config.theme.custom.is_none());
     }
 }
